Refiled XRGraphicsConfig and protected editor-only sections to enable building
