+++ com.unity.render-pipelines.core/CoreRP/Common/XRGraphicsConfig.cs.meta
