--- com.unity.render-pipelines.core/CoreRP/XR.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3f48c5fcf83adee4e815baff4b216f60
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
